Migrate ssh key generation to use ecdsa if the ssh-keygen binary is not available
This repo used to use rsa keys on windows and ed25519 keys on not-windows. This behavior was
controlled by compile time flags. This was non-optimal for two reasons. One, it used rsa keys
which are not preferred. Two, it meant that windows users who had the ssh-keygen binary
installed were not getting the benefits of using ed25519 keys even though their system does
support generating them. Now it is determined at run time whether to use ssh-keygen to generate
an ed25519 key or to generate an ecdsa key in pure go (without a dependency on ssh-keygen).
