+++ tests/tests/test_env_1.py
@@ -121,7 +121,7 @@ class TestMultiTeamStrictLogging:
                 if "----" in line and "PRIVATE" in line and "BEGIN" in line:
                     add = True
                 if add:
-                    keyLines.append(line)
+                    keyLines.append(line.strip())
                 if "----" in line and "PRIVATE" in line and "END" in line:
                     add = False
         key = '\n'.join(keyLines)
